requirements.txt
PyMySQL
python-dateutil